+++ packages/swapkit-api/src/api.ts
@@ -1,17 +1,19 @@
-import { QuoteParams, QuoteResponse } from './types.js';
+import { GasRatesResponse, QuoteParams, QuoteResponse } from './types/index.js';
 
-const baseUrl = `https://api.thorswap.net/aggregator`;
+const baseUrl = `https://api.thorswap.net`;
 const paramsToString = (params: any) => new URLSearchParams(params).toString();
 
 enum ApiEndpoints {
-  Quote = '/tokens/quote',
+  Quote = '/aggregator/tokens/quote',
+  GasRates = '/resource-worker/gasPrice/getAll',
 }
 
-const fetchWrapper = <T>(url: string, params?: any) =>
+const fetchWrapper = <T>(url: ApiEndpoints, params?: any) =>
   fetch(`${baseUrl}${url}${params ? `?${paramsToString(params)}` : ''}`).then(
     (res) => res.json() as Promise<T>,
   );
 
 export const SwapKitApi = {
   getQuote: (params: QuoteParams) => fetchWrapper<QuoteResponse>(ApiEndpoints.Quote, params),
+  getGasRates: () => fetchWrapper<GasRatesResponse>(ApiEndpoints.GasRates),
 };